refactor(engine): split SqliteVecStore out of embedding.rs into embedding/vec_store.rs
Pure structural refactor — zero behaviour change. Git detects the move
of embedding.rs → embedding/mod.rs as a rename (R100), so history is
preserved end-to-end.

Why:

embedding.rs was the single largest complexity pocket in the engine
crate at 2,934 LOC. Before the v1.5 Phase 1 split it held three
unrelated concerns in one file:

1. SqliteVecStore — SQLite + sqlite-vec backed EmbeddingStore impl
2. EmbeddingEngine facade + TextEmbeddingCache + helpers + model
   loading + the four analysis methods (classify / duplicates /
   similar / misplaced)
3. 670 LOC of tests touching both concerns

This commit extracts just concern #1 as the first of several planned
splits. The remaining mod.rs drops from 2,934 → 2,501 LOC. The new
vec_store.rs is a 451 LOC submodule with a crisp boundary: it owns
the persistence schema, the INSERT/SEARCH/DELETE statements, and the
`EmbeddingStore` trait impl. Nothing else.

What moved:

- `struct SqliteVecStore` (previously private)           → vec_store.rs
- `impl SqliteVecStore { new, insert_batch,
   decode_embedding_bytes, chunk_from_row }`             → vec_store.rs
- `impl EmbeddingStore for SqliteVecStore { upsert,
   insert, search, delete_by_file, clear, count,
   get_embedding, embeddings_for_scored_chunks,
   all_with_embeddings, embeddings_for_files,
   for_each_file_embeddings, for_each_embedding_batch }` → vec_store.rs

What stayed (visibility adjusted):

- `mod ffi { register_sqlite_vec, sysctl_usize }` now `pub(super)`
  (still used by `SqliteVecStore::new` and by macOS CPU detection
  in the engine facade — cannot move cleanly to one side)
- `fn embedding_to_bytes` now `pub(super)` (used by vec_store and
  by the test module — again shared, so it stays in mod.rs)

The `super::{embedding_to_bytes, ffi}` import in vec_store.rs is the
only place these leak across the new boundary. Everything else is
pure, self-contained, and unit-testable.

Follow-ups (not in this commit):

- Phase 1b: extract EmbeddingEngine + TextEmbeddingCache + model
  loading into embedding/engine.rs (~1500 LOC)
- Phase 1c: extract analysis result types + impl methods
  (classify_symbol, find_misplaced_code, find_similar_code,
  find_duplicates) into embedding/analysis.rs
- Phase 1d: split the 670 LOC test module into embedding/tests.rs

Each follow-up keeps tests green + clippy clean. They are sequenced
as separate commits so reviewers can bisect any behaviour change to
a ~400-LOC diff.

Verification:

  cargo test -p codelens-engine                  → 222 passed
  cargo test -p codelens-mcp                     → 148 passed
  cargo clippy --all-targets -- -D warnings      → 0 warnings
  cargo clippy --all-targets --features http     → 0 warnings
